v0.3 - patch 7.4 release candidate

+++ Ktisis/Editor/Animation/Game/ActionAnimation.cs
@@ -6,7 +6,7 @@ namespace Ktisis.Editor.Animation.Game;
 
 public class ActionAnimation(Action action) : GameAnimation {
 	public override string Name => action.Name.ExtractText();
-	public override ushort Icon => action.Icon;
+	public override uint Icon => action.Icon;
 	public override uint TimelineId => action.AnimationEnd.IsValid ? action.AnimationEnd.RowId : 0;
 	public override TimelineSlot Slot => (TimelineSlot)(action.AnimationEnd.IsValid ? action.AnimationEnd.Value.Stance : 0);
 }
